Navigation between pages
Set up an event listener to determine if a country is clicked and if it
is, build that page with the country data(not yet complete) and add it
to a list of visited pages so the HTML will not need to be built again
if the user revisits the same page in a single session. Also goes back
to the home page on Home or logo click. Added a variable to determine
which page the user is currently on so there will not be a refresh if
Home or logo is clicked while on the home page.